refactor: replace mail-type-detector and mail-time-detector with mail-helper crate
