+++ crates/rokbattles-api/Cargo.toml
@@ -12,12 +12,11 @@ tracing-subscriber = { workspace = true }
 futures-util = { workspace = true }
 mail-decoder = { path = "../mail-decoder" }
 mongodb = { workspace = true }
-mail-type-detector = { path = "../mail-type-detector" }
 zstd = { workspace = true }
 serde_json = { workspace = true }
 anyhow = { workspace = true }
 blake3 = { workspace = true }
-mail-time-detector = { path = "../mail-time-detector" }
+mail-helper = { path = "../mail-helper" }
 serde = { workspace = true }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
